Move rename refactoring classes to sub-package
