+++ plugins/org.eclipse.fordiac.ide.typemanagement/src/org/eclipse/fordiac/ide/typemanagement/refactoring/rename/RenameTypeRefactoringParticipant.java
@@ -11,7 +11,7 @@
  *   Fabio Gandolfi, Michael Oberlehner -
  *   	initial API and implementation and/or initial documentation
  *******************************************************************************/
-package org.eclipse.fordiac.ide.typemanagement.refactoring;
+package org.eclipse.fordiac.ide.typemanagement.refactoring.rename;
 
 import java.text.MessageFormat;
 import java.util.HashSet;
@@ -38,6 +38,10 @@ import org.eclipse.fordiac.ide.model.typelibrary.DataTypeEntry;
 import org.eclipse.fordiac.ide.model.typelibrary.TypeEntry;
 import org.eclipse.fordiac.ide.model.typelibrary.TypeLibraryManager;
 import org.eclipse.fordiac.ide.typemanagement.Messages;
+import org.eclipse.fordiac.ide.typemanagement.refactoring.InterfaceDataTypeChange;
+import org.eclipse.fordiac.ide.typemanagement.refactoring.StructuredTypeMemberChange;
+import org.eclipse.fordiac.ide.typemanagement.refactoring.UpdateInstancesChange;
+import org.eclipse.fordiac.ide.typemanagement.refactoring.UpdateTypeEntryChange;
 import org.eclipse.ltk.core.refactoring.Change;
 import org.eclipse.ltk.core.refactoring.CompositeChange;
 import org.eclipse.ltk.core.refactoring.RefactoringStatus;
@@ -46,10 +50,10 @@ import org.eclipse.ltk.core.refactoring.participants.RenameParticipant;
 
 public class RenameTypeRefactoringParticipant extends RenameParticipant {
 
-	IFile file;
-	TypeEntry typeEntry;
-	String oldName;
-	String newName;
+	private IFile file;
+	private TypeEntry typeEntry;
+	private String oldName;
+	private String newName;
 
 	@Override
 	protected boolean initialize(final Object element) {
